+++ example/lib/main.dart
@@ -1,6 +1,7 @@
 import 'dart:math';
 
 import 'package:calendar_day_view/calendar_day_view.dart';
+import 'package:english_words/english_words.dart';
 import 'package:example/tabs/event_day_view_tab.dart';
 import 'package:example/tabs/in_row_day_view_tab.dart';
 import 'package:example/tabs/overflow_day_view_tab.dart';
@@ -20,7 +21,7 @@ class MyApp extends StatelessWidget {
       debugShowCheckedModeBanner: false,
       title: 'Flutter Demo',
       theme: ThemeData(
-        primarySwatch: Colors.blue,
+        primarySwatch: Colors.teal,
       ),
       home: const CalendarDayViewExample(),
     );
@@ -36,13 +37,14 @@ class CalendarDayViewExample extends StatelessWidget {
       length: 3,
       child: SafeArea(
         child: Scaffold(
+          backgroundColor: Theme.of(context).backgroundColor,
           appBar: AppBar(
             title: const Text('Calendar Day Views'),
             bottom: const TabBar(
               tabs: [
+                Tab(text: 'Overflow'),
                 Tab(text: 'In Row'),
                 Tab(text: 'Event Only'),
-                Tab(text: 'Overflow'),
               ],
             ),
           ),
@@ -50,10 +52,6 @@ class CalendarDayViewExample extends StatelessWidget {
             padding: const EdgeInsets.symmetric(horizontal: 10),
             child: TabBarView(
               children: [
-                InRowDayViewTab(
-                  events: fakeEvents,
-                ),
-                EventDayViewTab(events: fakeEvents),
                 OverflowDayViewTab(
                   events: fakeEvents
                       .map(
@@ -66,6 +64,10 @@ class CalendarDayViewExample extends StatelessWidget {
                       )
                       .toList(),
                 ),
+                InRowDayViewTab(
+                  events: fakeEvents,
+                ),
+                EventDayViewTab(events: fakeEvents),
               ],
             ),
           ),
@@ -80,71 +82,71 @@ final List<DayEvent<String>> fakeEvents = [
   DayEvent(
     start: TimeOfDay.fromDateTime(
         DateTime(now.year, now.month, now.day, now.hour - 4)),
-    value: UniqueKey().toString(),
+    value: nouns.elementAt(rd.nextInt(2000)),
   ),
   DayEvent(
     start: TimeOfDay.fromDateTime(
         DateTime(now.year, now.month, now.day, now.hour + 3)),
-    value: UniqueKey().toString(),
+    value: nouns.elementAt(rd.nextInt(2000)),
   ),
   DayEvent(
     start: TimeOfDay.fromDateTime(
         DateTime(now.year, now.month, now.day, now.hour + 4)),
-    value: UniqueKey().toString(),
+    value: nouns.elementAt(rd.nextInt(2000)),
   ),
   DayEvent(
     start: TimeOfDay.fromDateTime(
         DateTime(now.year, now.month, now.day, now.hour - 2)),
-    value: UniqueKey().toString(),
+    value: nouns.elementAt(rd.nextInt(2000)),
   ),
   DayEvent(
     start: TimeOfDay.fromDateTime(
         DateTime(now.year, now.month, now.day, now.hour - 4)),
-    value: UniqueKey().toString(),
+    value: nouns.elementAt(rd.nextInt(2000)),
   ),
   DayEvent(
     start: TimeOfDay.fromDateTime(
         DateTime(now.year, now.month, now.day, now.hour - 4)),
-    value: UniqueKey().toString(),
+    value: nouns.elementAt(rd.nextInt(2000)),
   ),
   DayEvent(
     start: TimeOfDay.fromDateTime(
         DateTime(now.year, now.month, now.day, now.hour - 4)),
-    value: UniqueKey().toString(),
+    value: nouns.elementAt(rd.nextInt(2000)),
   ),
   DayEvent(
     start: TimeOfDay.fromDateTime(
         DateTime(now.year, now.month, now.day, now.hour - 4)),
-    value: UniqueKey().toString(),
+    value: nouns.elementAt(rd.nextInt(2000)),
   ),
   DayEvent(
     start: TimeOfDay.fromDateTime(
         DateTime(now.year, now.month, now.day, now.hour - 4)),
-    value: UniqueKey().toString(),
+    value: nouns.elementAt(rd.nextInt(2000)),
   ),
   DayEvent(
     start: TimeOfDay.fromDateTime(
         DateTime(now.year, now.month, now.day, now.hour - 3)),
-    value: UniqueKey().toString(),
+    value: nouns.elementAt(rd.nextInt(2000)),
   ),
   DayEvent(
     start: TimeOfDay.fromDateTime(
         DateTime(now.year, now.month, now.day, now.hour - 4, now.minute - 15)),
-    value: UniqueKey().toString(),
+    value: nouns.elementAt(rd.nextInt(2000)),
   ),
   DayEvent(
     start: TimeOfDay.fromDateTime(
         DateTime(now.year, now.month, now.day, now.hour - 4, now.minute - 15)),
-    value: UniqueKey().toString(),
+    value: nouns.elementAt(rd.nextInt(2000)),
   ),
   DayEvent(
     start: TimeOfDay.fromDateTime(
         DateTime(now.year, now.month, now.day, now.hour - 4, now.minute - 30)),
-    value: UniqueKey().toString(),
+    value: nouns.elementAt(rd.nextInt(2000)),
   ),
   DayEvent(
     start: TimeOfDay.fromDateTime(
         DateTime(now.year, now.month, now.day, now.hour - 4, now.minute - 30)),
-    value: UniqueKey().toString(),
+    value: nouns.elementAt(rd.nextInt(2000)),
   ),
 ];